openAIImageGenerator/requirements.txt
# DALL-E Image Generator Dependencies

# Core dependencies
pillow>=9.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# Utilities
requests>=2.27.0

# Testing
pytest>=7.0.0

# Additional dependencies
Pillow>=10.0.0
python-dateutil>=2.8.2 
